requirements.txt
pandas
scikit-learn
kafka-python 
mysql-connector-python 